fix: logic revised
