+++ app/(anon)/components/Input.tsx
@@ -8,6 +8,8 @@ interface InputProps {
   onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
   icon?: React.ReactNode;
   showPassword?: () => void;
+  disabled?: boolean;
+  required?: boolean;
 }
 export default function Input({
   type,
@@ -17,17 +19,22 @@ export default function Input({
   onChange,
   icon,
   showPassword,
+  disabled,
+  required = true,
 }: InputProps) {
   return (
     <div className="w-full flex py-[12px] px-[16px] bg-[#F8FAFC] justify-center rounded-[8px] border border-[#CBD5E1]">
       <input
-        required
+        required={required}
         type={type}
         name={name}
         placeholder={placeholder}
         value={value}
         onChange={onChange}
-        className="focus:outline-none bg-[#F8FAFC] text-[16px] font-normal text-[#1E293B] placeholder:text-[#94A3B8] w-full"
+        className={`focus:outline-none bg-[#F8FAFC] text-[16px] font-normal text-[#1E293B] placeholder:text-[#94A3B8] w-full ${
+          disabled ? 'bg-[#E2E8F0] text-[#94A3B8]' : ''
+        }`}
+        disabled={disabled}
       />
       {icon && <div onClick={showPassword}>{icon}</div>}
     </div>
