+++ backend/application/users/dtos/UpdateUserRequestDto.ts
@@ -1,7 +1,6 @@
 export interface UpdateUserRequestDto {
   id: number;
-  username: string;
-  email: string;
-  password: string;
-  nickname: string;
+  email?: string;
+  password?: string;
+  nickname?: string;
 }
